initial code cleaning ble code
