--- apps/AppWithWearable/lib/utils/actions/index.dart
@@ -1,9 +1,0 @@
-export 'ble0connect_device.dart' show ble0connectDevice;
-export 'ble0disconnect_device.dart' show ble0disconnectDevice;
-export 'ble0find_devices.dart' show ble0findDevices;
-export 'ble0get_connected_devices.dart' show ble0getConnectedDevices;
-export 'ble0get_rssi.dart' show ble0getRssi;
-export 'ble0receive_data.dart' show ble0receiveData;
-export 'ble0send_data.dart' show ble0sendData;
-export 'ble_receive_w_a_v.dart' show bleReceiveWAV;
-export 'ble_receive_wav_test.dart' show bleReceiveWavTest;
