+++ apps/AppWithWearable/lib/utils/ble/communication.dart
@@ -1,12 +1,8 @@
-// Automatic FlutterFlow imports
-// Begin custom action code
-// DO NOT REMOVE OR MODIFY THE CODE ABOVE!
-
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 
 import '/backend/schema/structs/index.dart';
 
-Future<String?> ble0receiveData(BTDeviceStruct btdevice) async {
+Future<String?> bleReceiveData(BTDeviceStruct btdevice) async {
   final device = BluetoothDevice.fromId(btdevice.id);
   final services = await device.discoverServices();
   for (BluetoothService service in services) {
@@ -20,4 +16,17 @@ Future<String?> ble0receiveData(BTDeviceStruct btdevice) async {
     }
   }
   return null;
+}
+
+Future bleSendData(BTDeviceStruct btDevice, String data) async {
+  final device = BluetoothDevice.fromId(btDevice.id);
+  final services = await device.discoverServices();
+  for (BluetoothService service in services) {
+    for (BluetoothCharacteristic characteristic in service.characteristics) {
+      final isWrite = characteristic.properties.write;
+      if (isWrite) {
+        await characteristic.write(data.codeUnits);
+      }
+    }
+  }
 }
