+++ apps/AppWithWearable/lib/utils/ble/connect.dart
@@ -1,18 +1,13 @@
-// Automatic FlutterFlow imports
 import 'package:flutter/material.dart';
-// Begin custom action code
-// DO NOT REMOVE OR MODIFY THE CODE ABOVE!
-
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
+import 'package:friend_private/backend/schema/structs/b_t_device_struct.dart';
 
-import '/backend/schema/structs/index.dart';
-
-Future<bool> ble0connectDevice(BTDeviceStruct btdevice) async {
-  final device = BluetoothDevice.fromId(btdevice.id);
+Future<bool> bleConnectDevice(BTDeviceStruct btDevice) async {
+  final device = BluetoothDevice.fromId(btDevice.id);
   try {
     await device.connect();
   } catch (e) {
-    print(e);
+    debugPrint(e.toString());
   }
   var hasWriteCharacteristic = false;
   final services = await device.discoverServices();
@@ -20,11 +15,19 @@ Future<bool> ble0connectDevice(BTDeviceStruct btdevice) async {
     for (BluetoothCharacteristic characteristic in service.characteristics) {
       final isWrite = characteristic.properties.write;
       if (isWrite) {
-        debugPrint(
-            'Found write characteristic: ${characteristic.uuid}, ${characteristic.properties}');
+        debugPrint('Found write characteristic: ${characteristic.uuid}, ${characteristic.properties}');
         hasWriteCharacteristic = true;
       }
     }
   }
   return hasWriteCharacteristic;
+}
+
+Future bleDisconnectDevice(BTDeviceStruct btDevice) async {
+  final device = BluetoothDevice.fromId(btDevice.id);
+  try {
+    await device.disconnect();
+  } catch (e) {
+    debugPrint(e.toString());
+  }
 }
