+++ apps/AppWithWearable/lib/utils/ble/find.dart
@@ -1,12 +1,8 @@
-// Automatic FlutterFlow imports
-// Begin custom action code
-// DO NOT REMOVE OR MODIFY THE CODE ABOVE!
-
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 
 import '/backend/schema/structs/index.dart';
 
-Future<List<BTDeviceStruct>> ble0findDevices() async {
+Future<List<BTDeviceStruct>> bleFindDevices() async {
   List<BTDeviceStruct> devices = [];
   FlutterBluePlus.scanResults.listen((results) {
     List<ScanResult> scannedDevices = [];
@@ -17,13 +13,13 @@ Future<List<BTDeviceStruct>> ble0findDevices() async {
     }
     scannedDevices.sort((a, b) => b.rssi.compareTo(a.rssi));
     devices.clear();
-    scannedDevices.forEach((deviceResult) {
+    for (var deviceResult in scannedDevices) {
       devices.add(BTDeviceStruct(
         name: deviceResult.device.name,
         id: deviceResult.device.id.toString(),
         rssi: deviceResult.rssi,
       ));
-    });
+    }
   });
   final isScanning = FlutterBluePlus.isScanningNow;
   if (!isScanning) {
