+++ apps/AppWithWearable/lib/utils/ble/scan.dart
@@ -1,14 +1,18 @@
 import 'package:friend_private/backend/schema/structs/b_t_device_struct.dart';
-import '/utils/actions/index.dart' as actions;
+import 'package:friend_private/utils/ble/connect.dart';
+import 'package:friend_private/utils/ble/find.dart';
 
 Future<BTDeviceStruct?> scanAndConnectDevice() async {
   while (true) {
-    List<BTDeviceStruct> foundDevices = await actions.ble0findDevices();
+    List<BTDeviceStruct> foundDevices = await bleFindDevices();
+    foundDevices.forEach((element) {
+      print(element);
+    });
     try {
       final friendDevice = foundDevices.firstWhere(
         (device) => device.name == 'Friend' || device.name == 'Super',
       );
-      await actions.ble0connectDevice(friendDevice);
+      await bleConnectDevice(friendDevice);
       return friendDevice;
     } catch (e) {
       // debugPrint('No matching device found, continue scanning');
